fix problem with large file
